+++ SpaceBattle.Lib/DataType/Angle.cs
