+++ SpaceBattle.Lib/DataType/Vector.cs
