+++ SpaceBattle.Lib/Move/IMovable.cs
@@ -1,0 +1,14 @@
+namespace SpaceBattle.Lib;
+
+public interface IMovable
+{
+    public Vector Position 
+    {
+        get;
+        set;
+    }
+    public Vector Velocity
+    {
+        get;
+    }
+}
