+++ SpaceBattle.Lib/Rotate/IRotatable.cs
@@ -1,0 +1,14 @@
+namespace SpaceBattle.Lib;
+
+public interface IRotatable
+{
+    public Angle Angle 
+    {
+        get;
+        set;
+    }
+    public Angle AngleVelocity 
+    {
+        get;
+    }
+}
