+++ SpaceBattle.Lib/Rotate/Rotate.cs
